* moved service_discovery helper into lib until it becomes a real npm module * retrieve credentials from the VCAP_SERVICES and attach them to the servicediscovery datasource to be passed later on to the ServiceDiscovery object * got rid of serviceDiscovery.js and trigger registration only after the app has started - so no one tries to reach the service if not only listening on http #19
