+++ server/boot/200-servicediscovery.js
@@ -1,14 +1,40 @@
 // Licensed under the Apache License. See footer for details.
 var winston = require("winston");
 
-var servicediscovery = require("../serviceDiscovery")
+var ServicePublisher = require('../../lib/service_discovery').ServicePublisher;
 
 module.exports = function (app) {
 
   //Register this applications url to the Service Discovery service
-  if(app.dataSources.servicediscovery){
-    var serviceDiscoveryCredentials = app.dataSources.servicediscovery.settings;
-    servicediscovery.registerInstance(serviceDiscoveryCredentials.serviceName, serviceDiscoveryCredentials.serviceEndpoint )
+  if (app.dataSources.servicediscovery) {
+
+    // Register with the service discovery once the REST API is mounted
+    app.once("started", function () {
+      var serviceDiscoveryDatasource = app.dataSources.servicediscovery.settings;
+
+      winston.info("Registering", serviceDiscoveryDatasource.serviceEndpoint,
+        "under the name", serviceDiscoveryDatasource.serviceName);
+
+      var publisher = new ServicePublisher(true, serviceDiscoveryDatasource.credentials);
+      publisher.publishService(serviceDiscoveryDatasource.serviceName,
+        serviceDiscoveryDatasource.serviceEndpoint, {
+          "tags": serviceDiscoveryDatasource.tags || ["logistics-wizard", "database"],
+          "TTL": serviceDiscoveryDatasource.ttl || 300
+        });
+
+      publisher.on('registered', function (serviceName) {
+        winston.info(serviceName + " successfully registered with Service Discovery!");
+      });
+
+      publisher.on('error', function (serverResponse) {
+        winston.error("Server responded with: ", serverResponse);
+      });
+
+      publisher.on('expired', function (serviceName) {
+        winston.error(serviceName + " expired from Service Discovery registry");
+      });
+    });
+    
   } else {
     winston.info("Service Discovery not found, skipping registration")
   }
