+++ server/datasources.local.js
@@ -27,10 +27,14 @@ if (process.env.VCAP_SERVICES) {
       "max": Math.max(1, vcapServices.elephantsql[0].credentials.max_conns - 2) // leave some connections for the frontend
     };
   }
-  if (vcapServices.hasOwnProperty("service_discovery")  && process.env.VCAP_APPLICATION) {
+  if (vcapServices.hasOwnProperty("service_discovery") && process.env.VCAP_APPLICATION) {
     datasources.servicediscovery = {
       "serviceName": "lw-erp",
       "serviceEndpoint": JSON.parse(process.env.VCAP_APPLICATION)['application_uris'][0],
+      "credentials": {
+        "token": vcapServices.service_discovery[0].credentials.auth_token,
+        "url": vcapServices.service_discovery[0].credentials.url
+      }
     };
   }
 }
