+++ server/server.js
@@ -22,13 +22,13 @@ require("events").EventEmitter.prototype._maxListeners = 100;
 app.start = function () {
   // start the web server
   return app.listen(function () {
-    app.emit("started");
     var baseUrl = app.get("url").replace(/\/$/, "");
     winston.info("Web server listening at: %s", baseUrl);
     if (app.get("loopback-component-explorer")) {
       var explorerPath = app.get("loopback-component-explorer").mountPath;
       winston.info("Browse your REST API at %s%s", baseUrl, explorerPath);
     }
+    app.emit("started");
   });
 };
